+++ tests/client-middleware-cache.js
@@ -2,20 +2,12 @@
 import { expect, should } from "chai";
 import sinon from "sinon";
 import cacheManager from "cache-manager";
+import jwt from "jwt-simple";
 
 import ShipCache from "../src/ship-cache";
 import Middleware from "../src/middleware/client";
 
-class HullStub {
-  constructor() {
-    this.logger = {
-      info: console.log, //() {},
-      debug: console.log, //() {}
-    }
-  }
-  get() {}
-  put() {}
-}
+import HullStub from "./support/hull-stub";
 
 const reqStub = {
   query: {
@@ -60,6 +52,7 @@ describe("Client Middleware", () => {
           value: "test2"
         }
       };
+
       shipCache.set("ship_id", newShip)
         .then((arg) => {
           instance(reqStub, {}, () => {
@@ -85,7 +78,7 @@ describe("Client Middleware", () => {
     });
   });
 
-  it("should share the cache using the same adapter", function (done) {
+  it("should share the cache using the same adapter and namespace", function (done) {
     const cacheAdapter = cacheManager.caching({ store: "memory", max: 100, ttl: 1/*seconds*/ });
     const shipCache1 = new ShipCache(cacheAdapter);
     const shipCache2 = new ShipCache(cacheAdapter);
@@ -110,39 +103,6 @@ describe("Client Middleware", () => {
               instance2(reqStub, {}, () => {
                 expect(reqStub.hull.ship.private_settings.value).to.equal("test2");
                 expect(this.getStub.calledOnce).to.be.true;
-                done();
-              });
-            });
-          });
-      });
-    });
-  });
-
-  it("should shupport different namespaces when using the same adapter", function (done) {
-    const cacheAdapter = cacheManager.caching({ store: "memory", max: 100, ttl: 1/*seconds*/ });
-    const shipCache1 = new ShipCache(cacheAdapter, "namespace1");
-    const shipCache2 = new ShipCache(cacheAdapter, "namespace2");
-    const instance1 = Middleware(HullStub, { hostSecret: "secret", shipCache: shipCache1 });
-    const instance2 = Middleware(HullStub, { hostSecret: "secret", shipCache: shipCache2 });
-    instance1(reqStub, {}, (err) => {
-      expect(reqStub.hull.ship.private_settings.value).to.equal("test");
-      expect(this.getStub.calledOnce).to.be.true;
-      instance2(reqStub, {}, () => {
-        expect(reqStub.hull.ship.private_settings.value).to.equal("test1");
-        expect(this.getStub.calledTwice).to.be.true;
-        const newShip = {
-          private_settings: {
-            value: "test2"
-          }
-        };
-        shipCache1.set("ship_id", newShip)
-          .then((arg) => {
-            instance1(reqStub, {}, () => {
-              expect(reqStub.hull.ship.private_settings.value).to.equal("test2");
-              expect(this.getStub.calledTwice).to.be.true;
-              instance2(reqStub, {}, () => {
-                expect(reqStub.hull.ship.private_settings.value).to.equal("test1");
-                expect(this.getStub.calledTwice).to.be.true;
                 done();
               });
             });